create page component revamp
